+++ pyAgxArm/protocols/can_protocol/drivers/nero/default/driver.py
@@ -8,7 +8,7 @@ from ......utiles.numeric_codec import (
     NumericCodec as nc,
     RAD2DEG,
 )
-from ......utiles.vaildator import Validator
+from ......utiles.validator import Validator
 from ....msgs.nero.default import (
     ArmMsgModeCtrl,
     ArmMsgFeedbackJointStates,
