+++ pyAgxArm/utiles/validator.py
